backend/requirements.txt
starlette_exporter
sqlalchemy
pydantic
psycopg2
python-slugify
aiofiles
pillow
python-multipart
tortoise
requests
passlib
pyjwt
python-dotenv
bcrypt